2021.3.22

+++ C++/LinkList/LinkList_Practice/LinkList.h
@@ -2,21 +2,20 @@
 #include <stdlib.h>
 #include <iostream>
 using namespace std;
-//使C++程序可以调用C语言程序
-	//定义节点数据类型
+//定义节点数据类型
 struct ListNode {
-	int val;
+	float val;
 	ListNode* next;
-	ListNode(int x) :val(x), next(nullptr) {}
+	ListNode(float x) :val(x), next(nullptr) {}
 	ListNode() :val(0), next(nullptr) {}
-	ListNode(int x, ListNode* next) :val(x), next(next) {}
+	ListNode(float x, ListNode* next) :val(x), next(next) {}
 };
 //初始化链表
 struct ListNode* Init_LinkList();
-//在值为oldval的后面插入一个新的数据newval
-void InsertByValue_LinkList(struct ListNode* header,int oldval,int newval);
+//在值为oldval的位置插入一个新的数据newval
+void InsertByValue_LinkList(struct ListNode* header, float oldval, float newval);
 //删除值为val的节点
-void RemoveByValue_LinkList(struct ListNode* header, int delValue);
+void RemoveByValue_LinkList(struct ListNode* header, float delValue);
 //遍历链表
 void Foreach_LinkList(struct ListNode* header);
 //销毁链表
